refactoring moving back tracking algorithm
