+++ app/src/main/java/com/holokenmod/backtrack/MathDokuCageBackTrack.java
@@ -1,10 +1,11 @@
-package com.holokenmod.creation;
+package com.holokenmod.backtrack;
 
 import android.util.Log;
 
 import com.holokenmod.Grid;
 import com.holokenmod.GridCage;
 import com.holokenmod.GridCell;
+import com.holokenmod.creation.GridCageCreator;
 
 import java.util.ArrayList;
 import java.util.List;
